+++ resources/upgrade-notification.sh
@@ -21,9 +21,8 @@ if [ "2.176.1-1" == "$(printf "%s\\n2.176.1-1" "${TO_VERSION}" | sort | head -n1
 fi
 
 # print upgrade notice for jenkins based on java 11
-
-if [ "2.190.3-4" == "$(printf "%s\\n2.190.3-4" "${TO_VERSION}" | sort | head -n1)" ] && [ "2.190.3-4" != "$(printf "%s\\n2.190.3-4" "${FROM_VERSION}" | sort | head -n1)" ]; then
-   printf "\nYou are upgrading the Jenkins dogu to version 2.190.3-4 or higher. There have been changes to the installed jdk version. \\nSince 2.190.3-4, jenkins is based on the open jdk version 11. \nLegacy builds using the internal jdk of jenkins must switch to the configured jdk <openJDK-8>."
-   printf "\nFor more information see %s \n\n" "https://github.com/cloudogu/jenkins/releases/tag/v2.190.3-4"
+if [ "2.222.1-2" == "$(printf "%s\\n2.222.1-2" "${TO_VERSION}" | sort | head -n1)" ] && [ "2.222.1-2" != "$(printf "%s\\n2.222.1-2" "${FROM_VERSION}" | sort | head -n1)" ]; then
+   printf "\nYou are upgrading the Jenkins dogu to version 2.222.1-2 or higher. There have been changes to the installed jdk version. \\nSince 2.222.1-2, jenkins is based on OpenJDK version 11. \nLegacy builds using the internal jdk of jenkins must switch to the configured jdk \"OpenJDK-8\"."
+   printf "\nFor more information see %s \n\n" "https://github.com/cloudogu/jenkins/blob/develop/README.md#building-with-openjdk-8"
 
 fi